refactor(plugins): begin overhaul - separating plugin specs from actual configuration
