+++ lua/core/options.lua
@@ -45,6 +45,8 @@ o.splitbelow = true
 o.splitright = true
 o.statuscolumn = [[%!v:lua.require'core.util'.statuscolumn()]]
 o.swapfile = false
+o.timeout = true
+o.timeoutlen = 300
 o.tabstop = 4
 o.termguicolors = true
 o.undodir = os.getenv('HOME') .. '/.vim/undodir'
