--- lua/plugins/coding/vim-illuminate.lua
@@ -1,59 +1,0 @@
-return {
-  {
-    'RRethy/vim-illuminate',
-    event = 'LazyFile',
-    opts = {
-      providers = {
-        'lsp',
-        'treesitter',
-        'regex',
-      },
-      filetypes_denylist = {
-        'alpha',
-        'dirbuf',
-        'dirvish',
-        'fugitive',
-        'harpoon',
-        'lazy',
-        'markdown.cody_history',
-        'markdown.cody_prompt',
-        'Mason',
-        'neo-tree',
-        'noice',
-        'suitui',
-        'TelescopePrompt',
-        'Trouble',
-      },
-      delay = 200,
-      large_file_cutoff = 2000,
-      large_file_overrides = {
-        providers = { 'lsp' },
-      },
-    },
-    config = function(_, opts)
-      require('illuminate').configure(opts)
-
-      local function map(key, dir, buffer)
-        vim.keymap.set('n', key, function()
-          require('illuminate')['goto_' .. dir .. '_reference'](false)
-        end, { desc = dir:sub(1, 1):upper() .. dir:sub(2) .. ' Reference', buffer = buffer })
-      end
-
-      map(']]', 'next')
-      map('[[', 'prev')
-
-      -- also set it after loading ftplugins, since a lot overwrite [[ and ]]
-      vim.api.nvim_create_autocmd('FileType', {
-        callback = function()
-          local buffer = vim.api.nvim_get_current_buf()
-          map(']]', 'next', buffer)
-          map('[[', 'prev', buffer)
-        end,
-      })
-    end,
-    keys = {
-      { ']]', desc = 'Next Reference' },
-      { '[[', desc = 'Prev Reference' },
-    },
-  },
-}
