+++ lua/plugins/config/notify.lua
@@ -1,0 +1,13 @@
+return function ()
+local opts = {
+  timeout = 3000,
+  max_height = function()
+    return math.floor(vim.o.lines * 0.75)
+  end,
+  max_width = function()
+    return math.floor(vim.o.columns * 0.75)
+  end,
+}
+
+require('notify').setup(opts)
+end
