+++ lua/plugins/config/telescope.lua
@@ -1,0 +1,18 @@
+return function()
+  local opts = {
+    extensions = {
+      fzf = {
+        fuzzy = true, -- false will only do exact matching
+        override_generic_sorter = true, -- override the generic sorter
+        override_file_sorter = true, -- override the file sorter
+        case_mode = 'smart_case', -- or "ignore_case" or "respect_case"
+        -- the default case_mode is "smart_case"
+      },
+    },
+  }
+
+  require('telescope').setup({ opts })
+
+  require('telescope').load_extension('fzf')
+  require('telescope').load_extension('notify')
+end
