+++ lua/plugins/config/whichkey.lua
@@ -1,16 +1,6 @@
-return {
-  {
-    'folke/which-key.nvim',
-    event = 'VeryLazy',
-    init = function()
-      vim.o.timeout = true
-      vim.o.timeoutlen = 300
-    end,
-    opts = {
-      -- your configuration comes here
-      -- or leave it empty to use the default settings
-      -- refer to the configuration section below
-      icons = {
+return function ()
+  local opts = {
+    icons = {
         breadcrumb = '»', -- symbol used in the command line area that shows your active key combo
         separator = '➜', -- symbol used between a key and it's label
         group = '', -- symbol prepended to a group
@@ -23,6 +13,8 @@ return {
         winblend = 0, -- value between 0-100 0 for fully opaque and 100 for fully transparent
         zindex = 1000, -- positive value to position WhichKey above other floating windows.
       },
-    },
-  },
-}
+    }
+
+    require('which-key').setup(opts)
+  end
+
