--- lua/plugins/utils/telescope.lua
@@ -1,52 +1,0 @@
-local util = require('core.util')
-
-return {
-  {
-    'nvim-telescope/telescope.nvim',
-    dependencies = {
-      { 'nvim-lua/plenary.nvim' },
-      { 'nvim-telescope/telescope-fzf-native.nvim', build = 'make' },
-    },
-    cmd = 'Telescope',
-    config = function()
-      require('telescope').setup({
-        extensions = {
-          fzf = {
-            fuzzy = true, -- false will only do exact matching
-            override_generic_sorter = true, -- override the generic sorter
-            override_file_sorter = true, -- override the file sorter
-            case_mode = 'smart_case', -- or "ignore_case" or "respect_case"
-            -- the default case_mode is "smart_case"
-          },
-        },
-      })
-      -- Essential Telescope extensions
-      require('telescope').load_extension('fzf')
-      require('telescope').load_extension('notify')
-    end,
-    keys = {
-      -- General
-      { '<leader>:', '<cmd>Telescope command_history<cr>', desc = 'Command History' },
-      { '<leader>,', '<cmd>Telescope buffers show_all_buffers=true<cr>', desc = 'Switch buffer' },
-      { '<leader>/', util.telescope('live_grep'), desc = 'Grep (root dir)' },
-      -- Files
-      { '<leader>fb', '<cmd>Telescope buffers<cr>', desc = 'Buffers' },
-      { '<leader>ff', util.telescope('files'), desc = 'Find files (root dir)' },
-      { '<leader>fF', util.telescope('files', { cwd = vim.loop.cwd() }), desc = 'Find files (cwd)' },
-      { '<leader>fr', '<cmd>Telescope oldfiles<cr>', desc = 'Recent files' },
-      { '<leader>fR', util.telescope('oldfiles', { cwd = vim.loop.cwd() }), desc = 'Recent files (cwd)' },
-      -- search
-      { '<leader>t"', '<cmd>Telescope registers<cr>', desc = 'Registers' },
-      { '<leader>tk', '<cmd>Telescope keymaps<cr>', desc = 'Keymaps' },
-      {
-        '<leader>uC',
-        function()
-          vim.api.nvim_exec_autocmds('User', { pattern = 'ColorSchemeLoad' })
-          vim.cmd('Telescope colorscheme')
-        end,
-        desc = 'Switch colorscheme',
-        noremap = true,
-      },
-    },
-  },
-}
